+++ libs/database-entity/src/lib.rs
@@ -1,3 +1,2 @@
-pub mod chat;
 pub mod dto;
 mod util;
